Changing "received date" to "approval date"
Not sure if this is the correct place to change this (?) assuming any changes made to the page need to happen in the master branch?  

Right now on the main page "explore submissions" the date sort is pulling on the received date, it would be better if it was pulling from the date it was signed by the CRO (chief records officer).  Thanks! Kate